+++ src/network_ops_dashboard/urls.py
@@ -35,9 +35,6 @@ urlpatterns = [
     re_path(r'^dashboard/save-prefs/$', views.dashboard_save_prefs, name='dashboard_save_prefs'),
     re_path(r'^dashboard/toggle/email_processing/$', views.dashboard_toggle_email_processing, name='dashboard_toggle_email_processing'),
     re_path(r'^dashboard/set/email_time/$', views.dashboard_set_email_time, name='dashboard_set_email_time'),
-    re_path(r'^dashboard/asavpn/save/$', views.dashboard_asa_vpn_settings_save, name='dashboard_asa_vpn_settings_save'),
-    re_path(r'^dashboard/sdwan/save/$', views.dashboard_sdwan_settings_save, name='dashboard_sdwan_settings_save'),
-    re_path(r'^dashboard/pagerduty/save/$', views.dashboard_pagerduty_settings_save, name='dashboard_pagerduty_settings_save'),
     re_path(r'^public_scripts/$', views.public_scripts, name='public_scripts'),
     re_path(r'^protected_media/(?P<path>.*)', views.protected_media, name='protected_media'),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT) + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
@@ -56,4 +53,5 @@ urlpatterns += [
     re_path(r'^notices/ciscoadvisory/', include('network_ops_dashboard.notices.ciscoadvisory.urls')),
     re_path(r'^notices/svcactexpiry/', include('network_ops_dashboard.notices.svcactexpiry.urls')),
     re_path(r'^notices/pagerduty/', include('network_ops_dashboard.notices.pagerduty.urls')),
+    re_path(r'^notices/statseeker/', include('network_ops_dashboard.notices.statseeker.urls')),
 ]